Allow inline `include` expansion of templates/media.
We've updated our static and dynamic templates to now allow
expansion of includes, via:

This means that we can avoid requesting files like /css/style.css,
or /js/k.js, instead serving them inline.

Because we're using includes there is some overhead at parse-time,
but we get the flexibilty to continue to edit the CSS/JS in one
place, rather than having to embed the same content in every file.

We now score 100/100 in Googles page-speed test, which closes #1.